socail-login

+++ database/migrations/2022_12_06_175958_create_reactions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('offer_id');
-            $table->integer('type')->unsigned()->default(1);
+            $table->tinyInteger('type')->unsigned()->default(1);
             $table->timestamps();
         });
     }
